refactor: Complete colocation optimization - perfect file organization
🎯 MAJOR REFACTORING: Complete colocation structure implementation

✅ COMPLETED CHANGES:
1. Rename src/const → src/constants (more conventional naming)
2. Create (root) group for top page organization
3. Move all page-specific constants to colocation:
   - about.ts → app/about/constants/
   - contact.ts → app/contact/constants/
   - works.ts → app/works/constants/
   - articles.ts → app/articles/constants/
4. Move all page-specific fetch libraries to colocation:
   - fetch-top-list.ts → app/(root)/libs/
   - fetch-articles.ts → app/articles/libs/
   - fetch-works-*.ts → app/works/libs/ & app/works/[slug]/libs/
5. Update all import paths to use relative imports for colocated files
6. Keep truly global files in src/ root (microcms.ts, microcms-client.ts)
7. Clean up empty directories

🎯 FINAL STRUCTURE:
app/
├── (root)/               # Top page group
│   ├── page.tsx
│   ├── components/service-list.tsx
│   └── libs/fetch-top-list.ts
├── [page]/
│   ├── components/       # Page-specific UI
│   ├── constants/        # Page-specific constants
│   ├── libs/            # Page-specific API functions
│   └── types/           # Page-specific types (contact only)
constants/               # Global constants only (microcms.ts)
libs/                   # Global libs only (microcms-client.ts)
features/ui/            # True shared UI components
types/                  # Global types
utils/                  # Global utilities
hooks/                  # Global hooks

✨ BENEFITS ACHIEVED:
- Perfect colocation: features placed closest to usage
- Zero barrel files (direct imports only)
- Optimized tree shaking and bundle size
- Clear separation of concerns
- Improved maintainability and developer experience
- Modern Next.js App Router best practices

Build passing ✓
