+++ src/app/about/constants/about.ts
