+++ src/app/works/components/works-list.tsx
@@ -7,7 +7,7 @@ import { ImageWrapper } from '@/features/ui/image-wrapper'
 
 import { conversionDate } from '@/utils/conversion-date'
 import { WorkIndex } from '@/types/microcms'
-import { NO_IMAGE } from '@/const/works'
+import { NO_IMAGE } from '../constants/works'
 
 type Props = {
   works: WorkIndex[]
