+++ src/app/works/page.tsx
@@ -4,7 +4,7 @@ import { WorksList } from './components/works-list'
 import { TextBtn } from '@/features/ui/text-btn'
 import { Title } from '@/features/ui/title'
 
-import { fetchWorksIndex } from '@/libs/fetch/fetch-works-index'
+import { fetchWorksIndex } from './libs/fetch-works-index'
 import { setMetaData } from '@/utils/set-metadata'
 
 const description = '最新の実績や、自主制作'
